requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
opencv-python>=4.5.0
tensorboard>=2.5.0
tqdm>=4.60.0
imageio>=2.9.0
scikit-image>=0.18.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0
open3d>=0.13.0
PyYAML>=6.0
pytest>=6.2.0
kornia>=0.5.0
dotmap>=1.3.0
configargparse>=1.5.0
plotly>=5.0.0
wandb>=0.12.0  # Optional for experiment tracking 